Data attributes modified
Data attributes are now a printer/parser pair to accommedate the
tutorial system, which is layer on top of the base game and so doesn't
have access to the game state. Information about the game state will be
encoded into data attributes and parsed by the tutorial system.

DataAttribute tests added to ensure that the print/parse round trip
returns the same values we statrted with.

Typesafe tutorials deemed unworkable, will be moving forward with the
new `GuideE` method of running tutorials.